Cambio de tipo de dato de los id, la conexion con la bdd funcionó

+++ src/main/java/com/example/aereopuerto/controller/ClienteController.java
@@ -29,7 +29,7 @@ public class ClienteController {
             @ApiResponse(responseCode = "404", description = "Cliente no encontrado")
     })
     @GetMapping("/{id}")
-    public ResponseEntity<Cliente> obtenerCliente(@PathVariable Long id) {
+    public ResponseEntity<Cliente> obtenerCliente(@PathVariable Integer id) {
         return ResponseEntity.ok(clienteService.obtenerClientePorId(id));
     }
 
@@ -54,7 +54,7 @@ public class ClienteController {
             @ApiResponse(responseCode = "404", description = "Cliente no encontrado")
     })
     @PutMapping("/{id}")
-    public ResponseEntity<Cliente> actualizarCliente(@PathVariable Long id, @RequestBody Cliente cliente) {
+    public ResponseEntity<Cliente> actualizarCliente(@PathVariable Integer id, @RequestBody Cliente cliente) {
         cliente.setId(id);
         return ResponseEntity.ok(clienteService.crearOActualizarCliente(cliente));
     }
@@ -62,7 +62,7 @@ public class ClienteController {
     @Operation(summary = "Eliminar cliente", description = "Elimina un cliente del sistema.")
     @ApiResponse(responseCode = "204", description = "Cliente eliminado exitosamente")
     @DeleteMapping("/{id}")
-    public ResponseEntity<Void> eliminarCliente(@PathVariable Long id) {
+    public ResponseEntity<Void> eliminarCliente(@PathVariable Integer id) {
         clienteService.eliminarCliente(id);
         return ResponseEntity.noContent().build();
     }
